Moving classes into more descriptive packages.
