+++ src/main/java/se/femtearenan/mindmap/model/Bubble.java
@@ -1,11 +1,11 @@
-package model;
+package se.femtearenan.mindmap.model;
 
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Ellipse;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.shape.Shape;
-import shapes.Cloudy;
-import shapes.Spiky;
+import se.femtearenan.mindmap.utility.shapes.Cloudy;
+import se.femtearenan.mindmap.utility.shapes.Spiky;
 
 import java.io.Serializable;
