+++ src/main/java/se/femtearenan/mindmap/model/BubbleType.java
@@ -1,4 +1,4 @@
-package model;
+package se.femtearenan.mindmap.model;
 
 public enum BubbleType {
     ELLIPSE, RECTANGLE, CLOUD, SPIKY
