+++ src/main/java/se/femtearenan/mindmap/model/IdeaConnectionType.java
@@ -1,4 +1,4 @@
-package model;
+package se.femtearenan.mindmap.model;
 
 public enum IdeaConnectionType {
     BRANCH, EXPLANATION, NONE
