+++ src/main/java/se/femtearenan/mindmap/utility/IdeaTracker.java
@@ -1,8 +1,8 @@
-package controller;
+package se.femtearenan.mindmap.utility;
 
+import se.femtearenan.mindmap.ui.IdeaController;
 import javafx.geometry.Point2D;
-import model.Idea;
-import model.PointSer;
+import se.femtearenan.mindmap.model.Idea;
 
 import java.io.Serializable;
 import java.util.*;
@@ -40,7 +40,7 @@ public class IdeaTracker implements Serializable{
         paneSceneTrack.put(idea, pointSer);
     }
 
-    Map<Idea, PointSer> getPaneSceneTrack() {
+    public Map<Idea, PointSer> getPaneSceneTrack() {
         return paneSceneTrack;
     }
 }
