+++ src/main/java/se/femtearenan/mindmap/utility/SelectionState.java
@@ -1,4 +1,4 @@
-package controller;
+package se.femtearenan.mindmap.utility;
 
 public enum SelectionState {
     NONE, SELECT_PARENT, SELECT_ACQUAINTANCE, REMOVE_CONNECTION
