+++ src/main/java/se/femtearenan/mindmap/utility/SizeChoice.java
@@ -1,4 +1,4 @@
-package controller;
+package se.femtearenan.mindmap.utility;
 
 public enum SizeChoice {
     SMALL, MEDIUM, LARGE
